+++ modules-explanations/nlg-spec-transformer/src/main/scala/org/openprovenance/prov/scala/nlgspec_transformer/SpecLoader.scala
@@ -188,6 +188,7 @@ object specTypes {
       }
     }
   }
+
   trait HasComplements {
     val complements: Seq[Phrase]
     def expandComplements(element: PhraseElement): Unit = {